% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFDM.R
\name{AFDM}
\alias{AFDM}
\title{Initialize Method for AFDM}
\description{
Initialize Method for AFDM

Initialize Method for AFDM
}
\examples{

## ------------------------------------------------
## Method `AFDM$new`
## ------------------------------------------------

# To create an instance of the class
afdm <- AFDM$new()


## ------------------------------------------------
## Method `AFDM$fit_transform`
## ------------------------------------------------

# Example usage with automatic component selection
df <- data.frame(a = rnorm(100), b = rnorm(100))
transformed_data <- afdm$fit_transform(df)

# Example usage with specified number of components
transformed_data <- afdm$fit_transform(df, n_compon = 5)


## ------------------------------------------------
## Method `AFDM$get_attributes`
## ------------------------------------------------

# Assuming 'fit_transform' has been already used to fit the model:
famd_model <- afdm$get_attributes()


## ------------------------------------------------
## Method `AFDM$get_var_characteristics`
## ------------------------------------------------

# Assuming 'fit_transform' has been already used to fit the model:
var_characteristics <- afdm$get_var_characteristics()


## ------------------------------------------------
## Method `AFDM$coord_supplementaries`
## ------------------------------------------------

# Assuming 'fit_transform' has been already used to fit the model:
new_data <- data.frame(a = rnorm(50), b = rnorm(50))
projected_data <- afdm$coord_supplementaries(new_data)


## ------------------------------------------------
## Method `AFDM$plot_eigen`
## ------------------------------------------------

# Assuming 'fit_transform' has been already used to fit the model:
afdm$plot_eigen()



## ------------------------------------------------
## Method `AFDM$plot_diagonal_auto`
## ------------------------------------------------

# Assuming 'fit_transform' has been already used to fit the model:
afdm$plot_diagonal_auto()



## ------------------------------------------------
## Method `AFDM$plot_cumsumprop`
## ------------------------------------------------

# Assuming 'fit_transform' has been already used to fit the model:
afdm$plot_cumsumprop()


}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AFDM-new}{\code{AFDM$new()}}
\item \href{#method-AFDM-fit_transform}{\code{AFDM$fit_transform()}}
\item \href{#method-AFDM-get_attributes}{\code{AFDM$get_attributes()}}
\item \href{#method-AFDM-get_var_characteristics}{\code{AFDM$get_var_characteristics()}}
\item \href{#method-AFDM-coord_supplementaries}{\code{AFDM$coord_supplementaries()}}
\item \href{#method-AFDM-plot_eigen}{\code{AFDM$plot_eigen()}}
\item \href{#method-AFDM-plot_diagonal_auto}{\code{AFDM$plot_diagonal_auto()}}
\item \href{#method-AFDM-plot_cumsumprop}{\code{AFDM$plot_cumsumprop()}}
\item \href{#method-AFDM-clone}{\code{AFDM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AFDM-new"></a>}}
\if{latex}{\out{\hypertarget{method-AFDM-new}{}}}
\subsection{Method \code{new()}}{
This method initializes an object of AFDM. It sets up the necessary
environment and variables for the object to function correctly. This method is
automatically called when a new object of this class is created.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AFDM$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
In the context of this R6 class, the initialize method may set up fields
         related to the FAMD model or other relevant data structures. It is essential
         for preparing the object for subsequent methods like 'fit_transform',
         'get_attributes', etc.
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# To create an instance of the class
afdm <- AFDM$new()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AFDM-fit_transform"></a>}}
\if{latex}{\out{\hypertarget{method-AFDM-fit_transform}{}}}
\subsection{Method \code{fit_transform()}}{
This function performs a Factor Analysis of Mixed Data (FAMD) on a given dataset.
It either automatically determines the number of components to keep based on the
elbow method if 'n_compon' is not specified, or uses the specified number of components.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AFDM$fit_transform(X_df, n_compon = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_df}}{A dataframe containing the data to be analyzed.}

\item{\code{n_compon}}{An optional integer specifying the number of components to retain.
If NULL, the function will automatically determine this number using the elbow method.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
If 'n_compon' is NULL, the function performs FAMD on the entire dataset
         and uses the elbow method to determine the optimal number of components
         to retain. The elbow method involves finding the point where the
         addition of another component doesn't significantly explain more variance.
         This is done by calculating the maximum vertical distance between the eigen values curve
         and a straight line linking the highest and smallest eigen values.
         If 'n_compon' is specified, the function retains the specified number
         of components. The function always returns the individual coordinates
         based on the selected components.
}

\subsection{Returns}{
A dataframe containing the coordinates of individuals on the retained components.
        The number of columns corresponds to the number of components kept.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Example usage with automatic component selection
df <- data.frame(a = rnorm(100), b = rnorm(100))
transformed_data <- afdm$fit_transform(df)

# Example usage with specified number of components
transformed_data <- afdm$fit_transform(df, n_compon = 5)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AFDM-get_attributes"></a>}}
\if{latex}{\out{\hypertarget{method-AFDM-get_attributes}{}}}
\subsection{Method \code{get_attributes()}}{
This function returns the attributes of the Factor Analysis of Mixed Data (FAMD) model
that has been fitted to the dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AFDM$get_attributes()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This function is useful for retrieving the complete FAMD model object,
         which includes various components like eigenvalues, scores, loadings, etc.
         This object can be used for further analysis or inspection of the model.
}

\subsection{Returns}{
Returns the FAMD model object with all its attributes.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Assuming 'fit_transform' has been already used to fit the model:
famd_model <- afdm$get_attributes()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AFDM-get_var_characteristics"></a>}}
\if{latex}{\out{\hypertarget{method-AFDM-get_var_characteristics}{}}}
\subsection{Method \code{get_var_characteristics()}}{
This method retrieves the coordinates and squared cosines (cos2) of the dimensions retained
for both qualitative and quantitative variables in the FAMD model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AFDM$get_var_characteristics(verbose = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{A logical value indicating whether to print the characteristics to the console.
Default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The method extracts the coordinates and cos2 from the FAMD model for both qualitative
         and quantitative variables. These characteristics are important for interpreting the
         factors in terms of the original variables. If `verbose` is TRUE, these details are
         printed to the console for easy viewing.
}

\subsection{Returns}{
A list containing coordinates and cos2 for qualitative and quantitative variables.
        The list has four elements: `coord_qual`, `cos2_qual`, `coord_quant`, and `cos2_quant`.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Assuming 'fit_transform' has been already used to fit the model:
var_characteristics <- afdm$get_var_characteristics()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AFDM-coord_supplementaries"></a>}}
\if{latex}{\out{\hypertarget{method-AFDM-coord_supplementaries}{}}}
\subsection{Method \code{coord_supplementaries()}}{
This function projects supplementary individuals (new observations) onto the factorial plane
defined by a previously fitted Factor Analysis of Mixed Data (FAMD) model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AFDM$coord_supplementaries(X_test)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_test}}{A dataframe containing the new observations to be projected.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function applies one-hot encoding to the test dataset and then
         scales it using the mean and standard deviation calculated from the training dataset.
         It then projects these scaled observations onto the factorial plane using the
         coefficients from the FAMD model.
}

\subsection{Returns}{
A dataframe containing the coordinates of the supplementary individuals
        in the space of the retained components of the FAMD model.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Assuming 'fit_transform' has been already used to fit the model:
new_data <- data.frame(a = rnorm(50), b = rnorm(50))
projected_data <- afdm$coord_supplementaries(new_data)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AFDM-plot_eigen"></a>}}
\if{latex}{\out{\hypertarget{method-AFDM-plot_eigen}{}}}
\subsection{Method \code{plot_eigen()}}{
This function plots a scree plot of eigenvalues for each factor/component in the FAMD model.
It helps in visualizing the importance of each factor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AFDM$plot_eigen()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The scree plot is a bar plot of eigenvalues of each factor in descending order.
         It is used to determine the number of factors to keep in factor analysis.
         The 'elbow' in the plot indicates the point after which the addition of
         new factors contributes less to explaining the variance in the data.
}

\subsection{Returns}{
A ggplot object representing the scree plot. The plot shows the eigenvalues
        of each factor as bars, making it easy to visualize the 'elbow'.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Assuming 'fit_transform' has been already used to fit the model:
afdm$plot_eigen()


}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AFDM-plot_diagonal_auto"></a>}}
\if{latex}{\out{\hypertarget{method-AFDM-plot_diagonal_auto}{}}}
\subsection{Method \code{plot_diagonal_auto()}}{
This function extends the basic scree plot by adding a diagonal line to assist
in the automatic detection of the 'elbow' point.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AFDM$plot_diagonal_auto()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This plot is similar to the basic scree plot but includes a diagonal line
         that helps in identifying the elbow point automatically. This is useful for
         determining the optimal number of factors to keep in an automated manner.
         The plot also includes labels showing the distance between the eigenvalues
         and the diagonal, providing a clear visual cue for the elbow.
}

\subsection{Returns}{
A ggplot object representing the scree plot with additional diagonal line.
        The plot shows the eigenvalues of each factor as bars and a diagonal line
        to facilitate the detection of the elbow point.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Assuming 'fit_transform' has been already used to fit the model:
afdm$plot_diagonal_auto()


}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AFDM-plot_cumsumprop"></a>}}
\if{latex}{\out{\hypertarget{method-AFDM-plot_cumsumprop}{}}}
\subsection{Method \code{plot_cumsumprop()}}{
This function creates a plot displaying the cumulative proportion of inertia explained by
each factor in the Factor Analysis of Mixed Data (FAMD) model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AFDM$plot_cumsumprop()}\if{html}{\out{</div>}}
}

\subsection{Details}{
The plot is particularly useful for assessing the contribution of each factor
         to the total variance explained.
}

\subsection{Returns}{
A ggplot object representing the cumulative proportions of inertia.
        The plot shows the cumulative proportion of inertia for each factor as bars,
        which helps in understanding the amount of variance explained by each factor.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{# Assuming 'fit_transform' has been already used to fit the model:
afdm$plot_cumsumprop()


}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AFDM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AFDM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AFDM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
