% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoricalNB.R
\name{CategoricalNB}
\alias{CategoricalNB}
\title{Categorical Naive Bayes Classifier}
\description{

An R6 objet of a Categorical Naive Bayes Classifier for qualitative and mixed (numerical and categorical) data.

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CategoricalNB-new}{\code{CategoricalNB$new()}}
\item \href{#method-CategoricalNB-preprocessing_train}{\code{CategoricalNB$preprocessing_train()}}
\item \href{#method-CategoricalNB-preprocessing_test}{\code{CategoricalNB$preprocessing_test()}}
\item \href{#method-CategoricalNB-fit}{\code{CategoricalNB$fit()}}
\item \href{#method-CategoricalNB-predict}{\code{CategoricalNB$predict()}}
\item \href{#method-CategoricalNB-predict_log_proba}{\code{CategoricalNB$predict_log_proba()}}
\item \href{#method-CategoricalNB-predict_proba}{\code{CategoricalNB$predict_proba()}}
\item \href{#method-CategoricalNB-print}{\code{CategoricalNB$print()}}
\item \href{#method-CategoricalNB-summary}{\code{CategoricalNB$summary()}}
\item \href{#method-CategoricalNB-score}{\code{CategoricalNB$score()}}
\item \href{#method-CategoricalNB-plot_confusionmatrix}{\code{CategoricalNB$plot_confusionmatrix()}}
\item \href{#method-CategoricalNB-plot_roccurve}{\code{CategoricalNB$plot_roccurve()}}
\item \href{#method-CategoricalNB-clone}{\code{CategoricalNB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-new"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$new(verbose = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-preprocessing_train"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-preprocessing_train}{}}}
\subsection{Method \code{preprocessing_train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$preprocessing_train(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The explanatory variables.}

\item{\code{y}}{The target variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The discretized dataframe.
Convert continuous data into discrete data for the predicted data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-preprocessing_test"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-preprocessing_test}{}}}
\subsection{Method \code{preprocessing_test()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$preprocessing_test(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The explanatory variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The discretized dataframe.
Fit the model according to the given training data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-fit"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$fit(X_train, y_train, lmbd = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_train}}{The explanatory variables.}

\item{\code{y_train}}{The target variables.}

\item{\code{lmbd}}{Optional constant used for Laplace smoothing.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The model.
Predict function.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-predict"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$predict(data_test, n_cores = 4)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_test}}{The dataset on which we wish to predict.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The prediction as a dataframe.
Predict log-probability estimates.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-predict_log_proba"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-predict_log_proba}{}}}
\subsection{Method \code{predict_log_proba()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$predict_log_proba(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The dataset on which we wish the probability prediction.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The dataframe which contains the log-probabilities.
Probability estimates.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$predict_proba(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The dataset on which we wish the probability prediction.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The dataframe which contains the probabilities.
Print function.
Summary function.
Evaluate Performance of Categorical Naive Bayes Classifier.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-print"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-summary"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-summary}{}}}
\subsection{Method \code{summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-score"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-score}{}}}
\subsection{Method \code{score()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$score(X_test, y_test, verbose = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_test}}{A dataframe containing the test set explanatory variables.}

\item{\code{y_test}}{A vector containing the actual class labels of the test set.}

\item{\code{verbose}}{A boolean indicating whether to print detailed output (default is TRUE).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the confusion matrix, error rate, 95% confidence interval for the error rate, recall, and precision for each class.
Visualization of the confusion matrix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-plot_confusionmatrix"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-plot_confusionmatrix}{}}}
\subsection{Method \code{plot_confusionmatrix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$plot_confusionmatrix(X, y_test)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The dataset X.}

\item{\code{y_test}}{y_test.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-plot_roccurve"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-plot_roccurve}{}}}
\subsection{Method \code{plot_roccurve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$plot_roccurve(X, y_test, positive = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CategoricalNB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CategoricalNB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CategoricalNB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
