% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{This function provides a summary for NaiveBayes R6 objects.}
\usage{
summary(nb, ...)
}
\arguments{
\item{nb}{An object of class GaussianNB, BernoulliNB, or CategoricalNB.}

\item{print_afdm}{A boolean to print the FAMD results in case the model is a Gaussian Naive Bayes Classifier on mixed data.}
}
\value{
Summary of the Naive Bayes Classifier object.
}
\description{
This function provides a summary for NaiveBayes R6 objects.
}
\details{
R6 does not have a summary generic method, so we have to add a small S3 method function to call the summary function from R6 object.
}
