% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\docType{data}
\name{text}
\alias{text}
\title{Text dataset with Binary Classification - Spam dataset}
\format{
A data frame of 5572 messages with the following columns:
\describe{
  \item{Class}{"spam", "ham"}
  \item{Message}{Message}
}
}
\source{
\url{https://www.kaggle.com/datasets/uciml/sms-spam-collection-dataset}
}
\usage{
data(text)
}
\description{
The SMS Spam Collection is a set of SMS tagged messages that have been collected for SMS Spam research.
It contains one set of SMS messages in English of 5572 messages, tagged acording being ham (legitimate)
or spam.
}
\details{
Almeida, T.A., Gomez Hidalgo, J.M., Yamakami, A. Contributions to the Study of SMS Spam Filtering: New Collection and Results.
Proceedings of the 2011 ACM Symposium on Document Engineering (DOCENG'11), Mountain View, CA, USA, 2011.
}
\keyword{datasets}
